pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brand"
version = "0.1.1"
description = "A composable pipeline for brand name generation, evaluation, and availability checking"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "Thor Whalen"},
]
keywords = ["brand", "naming", "domain", "availability", "pipeline", "phonetics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Internet :: Name Service (DNS)",
]

dependencies = [
    "dol",
    "config2py",
    "argh",
    "requests",
    "lexis",
    "python-whois",
    "timeout-decorator",
    "wordfreq",
    "pronouncing",
]

[project.optional-dependencies]
phonetics = [
    "epitran",
    "panphon",
    "python-BLICK",
]
all = [
    "epitran",
    "panphon",
    "python-BLICK",
    "oa",
]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1.0"]

[project.urls]
Homepage = "https://github.com/thorwhalen/brand"
Repository = "https://github.com/thorwhalen/brand"

[tool.hatch.build.targets.wheel]
packages = ["brand"]

[tool.hatch.build]
include = ["brand/**"]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    "**/*.ipynb",
    ".git",
    ".venv",
    "build",
    "dist",
    "tests",
    "examples",
    "scrap",
]

[tool.ruff.lint]
select = [
    "D100",
]
ignore = [
    "D203",
    "E501",
    "B905",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["D"]
"**/examples/*" = ["D"]
"**/scrap/*" = ["D"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
doctest_optionflags = ["NORMALIZE_WHITESPACE", "ELLIPSIS"]

[tool.wads.ci]
project_name = "brand"

[tool.wads.ci.testing]
python_versions = ["3.10", "3.12"]
pytest_args = ["-v", "--tb=short"]
coverage_enabled = true
exclude_paths = ["examples", "scrap"]
test_on_windows = true

[tool.wads.ci.build]
sdist = true
wheel = true
